QTribe_api/requirements.txt
# QTribe API Requirements
# Django REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
django-redis>=5.4.0

# JWT Authentication
djangorestframework-simplejwt>=5.3.0

# Database
mysqlclient>=2.2.0
PyMySQL>=1.1.0

# Redis
redis>=5.0.0

# Security
bleach>=6.1.0

# File handling
Pillow>=10.0.0

# Celery for async tasks
celery>=5.3.0
django-celery-results>=2.5.0

# Search
django-haystack>=3.2.1
elasticsearch>=7.17.0

# Third-party login
AgentLogin>=3.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Production server
gunicorn>=21.0.0
